Move PageRoutedTabs to the frontend as it is causing issues in the framework
