+++ frontend/eda/access/users/UserPage/MyPage.tsx
@@ -14,8 +14,8 @@ import {
   useGetPageUrl,
   usePageNavigate,
 } from '../../../../../framework';
-import { PageRoutedTabs } from '../../../../../framework/PageTabs/PageRoutedTabs';
 import { LoadingPage } from '../../../../../framework/components/LoadingPage';
+import { PageRoutedTabs } from '../../../../common/PageRoutedTabs';
 import { useGet } from '../../../../common/crud/useGet';
 import { edaAPI } from '../../../common/eda-utils';
 import { useEdaActiveUser } from '../../../common/useEdaActiveUser';
