+++ frontend/eda/event-streams/EventStreamInstancePage/EventStreamInstancePage.tsx
@@ -1,11 +1,11 @@
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import { PageHeader, PageLayout, useGetPageUrl } from '../../../../framework';
-import { PageRoutedTabs } from '../../../../framework/PageTabs/PageRoutedTabs';
+import { PageRoutedTabs } from '../../../common/PageRoutedTabs';
 import { useGet } from '../../../common/crud/useGet';
 import { edaAPI } from '../../common/eda-utils';
-import { EdaEventStreamInstance } from '../../interfaces/EdaEventStreamInstance';
 import { EdaEventStream } from '../../interfaces/EdaEventStream';
+import { EdaEventStreamInstance } from '../../interfaces/EdaEventStreamInstance';
 import { EdaRoute } from '../../main/EdaRoutes';
 
 export function EventStreamInstancePage() {
