+++ frontend/hub/execution-environments/ExecutionEnvironmentPage/ExecutionEnvironmentImagePage.tsx
@@ -1,21 +1,21 @@
+import { ClipboardCopyButton, LabelGroup, Stack, StackItem } from '@patternfly/react-core';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
-import { LabelGroup, ClipboardCopyButton, Stack, StackItem } from '@patternfly/react-core';
 import {
+  BytesCell,
   LoadingPage,
   PageHeader,
   PageLayout,
   useGetPageUrl,
-  BytesCell,
 } from '../../../../framework';
-import { PageRoutedTabs } from '../../../../framework/PageTabs/PageRoutedTabs';
+import { PageRoutedTabs } from '../../../common/PageRoutedTabs';
+import { useGet } from '../../../common/crud/useGet';
 import { HubError } from '../../common/HubError';
 import { hubAPI } from '../../common/api/formatPath';
-import { useGet } from '../../../common/crud/useGet';
 import { HubRoute } from '../../main/HubRoutes';
-import { ShaLabel, TagLabel } from './components/ImageLabels';
 import { ExecutionEnvironment } from '../ExecutionEnvironment';
-import { ImageLayer, ExecutionEnvironmentImage } from './ExecutionEnvironmentImage';
+import { ExecutionEnvironmentImage, ImageLayer } from './ExecutionEnvironmentImage';
+import { ShaLabel, TagLabel } from './components/ImageLabels';
 
 export function ExecutionEnvironmentImagePage() {
   const { t } = useTranslation();
